Split Vulkan kernel policy module
