+++ crates/kiln-vulkan-kernel/src/policy.rs
@@ -1,0 +1,318 @@
+/// Immutable, device-neutral kernel selection used by Vulkan execution.
+///
+/// Product defaults must never depend on device names, vendor IDs, or
+/// qualification-machine identities. Optional routes belong behind explicit
+/// Vulkan feature and limit checks.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct VulkanKernelPolicy {
+    pub flash_attention_row_tile: usize,
+    pub flash_attention_row_work_budget: usize,
+    pub bf16_weight_row_tile: usize,
+    pub elementwise_tile_elements: usize,
+    pub exp_tile_elements: usize,
+    pub gdn_enabled: bool,
+    pub gdn_prefill_in_proj_enabled: bool,
+    pub gdn_gates_enabled: bool,
+    pub gdn_gated_rms_norm_enabled: bool,
+    pub gdn_full_chunk_forward_enabled: bool,
+    pub fused_conv1d_update_enabled: bool,
+    pub fused_conv1d_prefill_enabled: bool,
+    pub conv1d_prefill_single_submit_enabled: bool,
+    pub gdn_forward_sub_enabled: bool,
+    pub gdn_decode_fused_enabled: bool,
+    pub gdn_recurrent_unexpanded_qk_enabled: bool,
+    pub gdn_recurrent_qk_norm_unexpanded_enabled: bool,
+    pub linear_decode_enabled: bool,
+    pub linear_argmax_batch_enabled: bool,
+    pub full_attn_qkv_enabled: bool,
+    pub paged_attn_decode_batch_enabled: bool,
+    pub mlp_decode_enabled: bool,
+    pub mlp_gate_up_enabled: bool,
+    pub mlp_bf16_gate_up_f32_down_enabled: bool,
+    pub bf16_packed_linear_weights_enabled: bool,
+    pub bf16_packed_gdn_in_proj_weights_enabled: bool,
+    pub bf16_packed_full_attn_qkv_weights_enabled: bool,
+    pub bf16_packed_mlp_decode_weights_enabled: bool,
+    pub recurrent_state_residency_enabled: bool,
+    pub prefill_recurrent_state_residency_enabled: bool,
+    pub resident_decode_enabled: bool,
+    pub bridged_rmsnorm_forward_enabled: bool,
+    pub skip_final_gdn_state_readback_enabled: bool,
+    pub flash_attn_prefill_enabled: bool,
+    pub paged_decode_gpu_gather_enabled: bool,
+    pub gdn_chunkwise_forward_enabled: bool,
+    pub gdn_chunkwise_single_submit_enabled: bool,
+    pub gdn_chunkwise_fallback_enabled: bool,
+    pub gdn_decode_fused_resident_state_enabled: bool,
+    pub linear_max_flop_per_dispatch: u64,
+    pub mlp_bf16_gate_up_rows4: bool,
+    pub mlp_f32_down_rows4: bool,
+    pub mlp_bf16_down_rows4: bool,
+    pub mlp_bf16_rows8: bool,
+    pub mlp_bf16_rows8_min_batch: usize,
+    pub mlp_bf16_gate_up_rows4_min_batch: usize,
+    pub mlp_bf16_down_rows4_min_batch: usize,
+    pub mlp_f32_down_rows4_min_batch: usize,
+    pub linear_decode_bf16w_rows4: bool,
+    pub linear_decode_bf16w_rows8: bool,
+    pub linear_bf16_rows4_min_batch: usize,
+    pub linear_bf16_rows8_min_batch: usize,
+    pub gdn_in_proj_rows4_min_batch: usize,
+    pub gdn_in_proj_rows8_min_batch: usize,
+    pub full_attn_qkv_bf16w_rows4: bool,
+    pub full_attn_qkv_bf16w_rows8: bool,
+    pub full_attn_qkv_bf16_rows4_min_batch: usize,
+    pub full_attn_qkv_bf16_rows8_min_batch: usize,
+    pub paged_attn_single_submit: bool,
+    pub qwen_rmsnorm_single_submit: bool,
+    pub gdn_gates_single_submit: bool,
+    pub gdn_gated_norm_single_submit: bool,
+    pub mlp_gate_up_single_submit: bool,
+    pub causal_conv1d_single_submit: bool,
+    pub mlp_chained_dispatch: bool,
+    pub mlp_chained_transfer_submit: bool,
+    pub gdn_decode_host_visible_state: bool,
+    pub gdn_decode_fused_single_submit: bool,
+    pub gdn_recurrent_host_visible_state: bool,
+    pub gdn_recurrent_host_visible_batch_state: bool,
+    pub gdn_recurrent_single_submit: bool,
+    pub gdn_recurrent_parallel_reduce: bool,
+    pub linear_decode_single_submit: bool,
+    pub linear_decode_argmax_single_submit: bool,
+    pub full_attn_qkv_single_submit: bool,
+    pub gdn_in_proj_single_submit: bool,
+    pub gdn_in_proj_batch_pair_qkv_z: bool,
+    pub gdn_in_proj_batch_row_pair: bool,
+    pub gdn_in_proj_batch_row_quad: bool,
+    pub gdn_in_proj_batch_row_octet: bool,
+    pub gdn_gates_batched_transfers: bool,
+    pub gdn_gated_norm_batched_uploads: bool,
+    pub gdn_chunk_batched_transfers: bool,
+    pub paged_attn_batched_uploads: bool,
+    pub prefill_row_pair_matmul: bool,
+    pub gdn_qk_norm_recurrent_fusion: bool,
+    pub gdn_in_proj_conv_split_fusion: bool,
+    pub profile_mlp_kernel_stages: bool,
+    pub profile_gdn_in_proj_kernel_stages: bool,
+    pub profile_gdn_recurrent_kernel_stages: bool,
+    pub profile_resident_decode_timing: bool,
+}
+
+pub const VULKAN_KERNEL_POLICY_SCHEMA_ID: &str = "kiln.vulkan-kernel-policy.v5";
+
+impl VulkanKernelPolicy {
+    /// Fast device-neutral routes implemented with the Vulkan 1.2 compute
+    /// contract used by Kiln's logical device.
+    ///
+    /// These shaders do not require a machine profile or vendor extension.
+    /// Individual dispatchers remain responsible for validating the limits
+    /// their shader and workload require.
+    pub const fn native_default() -> Self {
+        Self {
+            flash_attention_row_tile: 2048,
+            flash_attention_row_work_budget: 10_000_000,
+            bf16_weight_row_tile: 128,
+            elementwise_tile_elements: 1 << 20,
+            exp_tile_elements: 65_536,
+            gdn_enabled: true,
+            gdn_prefill_in_proj_enabled: true,
+            gdn_gates_enabled: true,
+            gdn_gated_rms_norm_enabled: true,
+            gdn_full_chunk_forward_enabled: false,
+            fused_conv1d_update_enabled: false,
+            fused_conv1d_prefill_enabled: true,
+            conv1d_prefill_single_submit_enabled: true,
+            gdn_forward_sub_enabled: false,
+            gdn_decode_fused_enabled: false,
+            gdn_recurrent_unexpanded_qk_enabled: true,
+            gdn_recurrent_qk_norm_unexpanded_enabled: true,
+            linear_decode_enabled: true,
+            linear_argmax_batch_enabled: true,
+            full_attn_qkv_enabled: true,
+            paged_attn_decode_batch_enabled: true,
+            mlp_decode_enabled: true,
+            mlp_gate_up_enabled: false,
+            mlp_bf16_gate_up_f32_down_enabled: true,
+            bf16_packed_linear_weights_enabled: true,
+            bf16_packed_gdn_in_proj_weights_enabled: true,
+            bf16_packed_full_attn_qkv_weights_enabled: true,
+            bf16_packed_mlp_decode_weights_enabled: true,
+            recurrent_state_residency_enabled: false,
+            prefill_recurrent_state_residency_enabled: false,
+            resident_decode_enabled: true,
+            bridged_rmsnorm_forward_enabled: false,
+            skip_final_gdn_state_readback_enabled: true,
+            flash_attn_prefill_enabled: true,
+            paged_decode_gpu_gather_enabled: true,
+            gdn_chunkwise_forward_enabled: true,
+            gdn_chunkwise_single_submit_enabled: true,
+            gdn_chunkwise_fallback_enabled: false,
+            gdn_decode_fused_resident_state_enabled: true,
+            linear_max_flop_per_dispatch: 20_000_000_000,
+            mlp_bf16_gate_up_rows4: true,
+            mlp_f32_down_rows4: true,
+            mlp_bf16_down_rows4: true,
+            mlp_bf16_rows8: true,
+            mlp_bf16_rows8_min_batch: 256,
+            mlp_bf16_gate_up_rows4_min_batch: 8,
+            mlp_bf16_down_rows4_min_batch: 16,
+            mlp_f32_down_rows4_min_batch: 8,
+            linear_decode_bf16w_rows4: true,
+            linear_decode_bf16w_rows8: true,
+            linear_bf16_rows4_min_batch: 16,
+            linear_bf16_rows8_min_batch: 64,
+            gdn_in_proj_rows4_min_batch: 16,
+            gdn_in_proj_rows8_min_batch: 64,
+            full_attn_qkv_bf16w_rows4: true,
+            full_attn_qkv_bf16w_rows8: true,
+            full_attn_qkv_bf16_rows4_min_batch: 2,
+            full_attn_qkv_bf16_rows8_min_batch: 64,
+            paged_attn_single_submit: true,
+            qwen_rmsnorm_single_submit: true,
+            gdn_gates_single_submit: true,
+            gdn_gated_norm_single_submit: true,
+            mlp_gate_up_single_submit: true,
+            causal_conv1d_single_submit: true,
+            mlp_chained_dispatch: true,
+            mlp_chained_transfer_submit: true,
+            gdn_decode_host_visible_state: false,
+            gdn_decode_fused_single_submit: false,
+            gdn_recurrent_host_visible_state: true,
+            gdn_recurrent_host_visible_batch_state: false,
+            gdn_recurrent_single_submit: true,
+            gdn_recurrent_parallel_reduce: true,
+            linear_decode_single_submit: true,
+            linear_decode_argmax_single_submit: true,
+            full_attn_qkv_single_submit: true,
+            gdn_in_proj_single_submit: true,
+            gdn_in_proj_batch_pair_qkv_z: true,
+            gdn_in_proj_batch_row_pair: true,
+            gdn_in_proj_batch_row_quad: true,
+            gdn_in_proj_batch_row_octet: false,
+            gdn_gates_batched_transfers: true,
+            gdn_gated_norm_batched_uploads: true,
+            gdn_chunk_batched_transfers: true,
+            paged_attn_batched_uploads: true,
+            prefill_row_pair_matmul: true,
+            gdn_qk_norm_recurrent_fusion: true,
+            gdn_in_proj_conv_split_fusion: false,
+            profile_mlp_kernel_stages: false,
+            profile_gdn_in_proj_kernel_stages: false,
+            profile_gdn_recurrent_kernel_stages: false,
+            profile_resident_decode_timing: false,
+        }
+    }
+
+    /// Decline optional optimized Vulkan routes while retaining bounded-work
+    /// limits and explicit reference fallbacks.
+    pub const fn portable_fallback() -> Self {
+        Self {
+            flash_attention_row_tile: 2048,
+            flash_attention_row_work_budget: 10_000_000,
+            bf16_weight_row_tile: 128,
+            elementwise_tile_elements: 1 << 20,
+            exp_tile_elements: 65_536,
+            gdn_enabled: false,
+            gdn_prefill_in_proj_enabled: false,
+            gdn_gates_enabled: false,
+            gdn_gated_rms_norm_enabled: false,
+            gdn_full_chunk_forward_enabled: false,
+            fused_conv1d_update_enabled: false,
+            fused_conv1d_prefill_enabled: false,
+            conv1d_prefill_single_submit_enabled: false,
+            gdn_forward_sub_enabled: false,
+            gdn_decode_fused_enabled: false,
+            gdn_recurrent_unexpanded_qk_enabled: false,
+            gdn_recurrent_qk_norm_unexpanded_enabled: false,
+            linear_decode_enabled: false,
+            linear_argmax_batch_enabled: false,
+            full_attn_qkv_enabled: false,
+            paged_attn_decode_batch_enabled: false,
+            mlp_decode_enabled: false,
+            mlp_gate_up_enabled: false,
+            mlp_bf16_gate_up_f32_down_enabled: false,
+            bf16_packed_linear_weights_enabled: false,
+            bf16_packed_gdn_in_proj_weights_enabled: false,
+            bf16_packed_full_attn_qkv_weights_enabled: false,
+            bf16_packed_mlp_decode_weights_enabled: false,
+            recurrent_state_residency_enabled: false,
+            prefill_recurrent_state_residency_enabled: false,
+            resident_decode_enabled: false,
+            bridged_rmsnorm_forward_enabled: false,
+            skip_final_gdn_state_readback_enabled: false,
+            flash_attn_prefill_enabled: false,
+            paged_decode_gpu_gather_enabled: false,
+            gdn_chunkwise_forward_enabled: false,
+            gdn_chunkwise_single_submit_enabled: false,
+            gdn_chunkwise_fallback_enabled: true,
+            gdn_decode_fused_resident_state_enabled: false,
+            linear_max_flop_per_dispatch: 20_000_000_000,
+            mlp_bf16_gate_up_rows4: false,
+            mlp_f32_down_rows4: false,
+            mlp_bf16_down_rows4: false,
+            mlp_bf16_rows8: false,
+            mlp_bf16_rows8_min_batch: 256,
+            mlp_bf16_gate_up_rows4_min_batch: 8,
+            mlp_bf16_down_rows4_min_batch: 16,
+            mlp_f32_down_rows4_min_batch: 8,
+            linear_decode_bf16w_rows4: false,
+            linear_decode_bf16w_rows8: false,
+            linear_bf16_rows4_min_batch: 16,
+            linear_bf16_rows8_min_batch: 64,
+            gdn_in_proj_rows4_min_batch: 16,
+            gdn_in_proj_rows8_min_batch: 64,
+            full_attn_qkv_bf16w_rows4: false,
+            full_attn_qkv_bf16w_rows8: false,
+            full_attn_qkv_bf16_rows4_min_batch: 2,
+            full_attn_qkv_bf16_rows8_min_batch: 64,
+            paged_attn_single_submit: false,
+            qwen_rmsnorm_single_submit: false,
+            gdn_gates_single_submit: false,
+            gdn_gated_norm_single_submit: false,
+            mlp_gate_up_single_submit: false,
+            causal_conv1d_single_submit: false,
+            mlp_chained_dispatch: false,
+            mlp_chained_transfer_submit: false,
+            gdn_decode_host_visible_state: false,
+            gdn_decode_fused_single_submit: false,
+            gdn_recurrent_host_visible_state: true,
+            gdn_recurrent_host_visible_batch_state: false,
+            gdn_recurrent_single_submit: false,
+            gdn_recurrent_parallel_reduce: false,
+            linear_decode_single_submit: false,
+            linear_decode_argmax_single_submit: false,
+            full_attn_qkv_single_submit: false,
+            gdn_in_proj_single_submit: false,
+            gdn_in_proj_batch_pair_qkv_z: false,
+            gdn_in_proj_batch_row_pair: false,
+            gdn_in_proj_batch_row_quad: false,
+            gdn_in_proj_batch_row_octet: false,
+            gdn_gates_batched_transfers: false,
+            gdn_gated_norm_batched_uploads: false,
+            gdn_chunk_batched_transfers: false,
+            paged_attn_batched_uploads: false,
+            prefill_row_pair_matmul: false,
+            gdn_qk_norm_recurrent_fusion: false,
+            gdn_in_proj_conv_split_fusion: false,
+            profile_mlp_kernel_stages: false,
+            profile_gdn_in_proj_kernel_stages: false,
+            profile_gdn_recurrent_kernel_stages: false,
+            profile_resident_decode_timing: false,
+        }
+    }
+}
+
+impl Default for VulkanKernelPolicy {
+    fn default() -> Self {
+        Self::native_default()
+    }
+}
+
+pub const NATIVE_VULKAN_KERNEL_POLICY: VulkanKernelPolicy = VulkanKernelPolicy::native_default();
+
+pub const PORTABLE_VULKAN_KERNEL_POLICY: VulkanKernelPolicy =
+    VulkanKernelPolicy::portable_fallback();
+
+pub fn vulkan_kernel_policy() -> VulkanKernelPolicy {
+    NATIVE_VULKAN_KERNEL_POLICY
+}
